+++ docs/utils.rst
@@ -1,4 +1,4 @@
-.. _containers:
+.. _transformations:
 
 Helper utilities
 ================
